refactor and reorganize
