--- format.sh
@@ -1,1 +1,0 @@
-cargo fmt
