+++ scripts/build.sh
@@ -1,1 +1,2 @@
+#!/bin/bash
 cargo build
