+++ scripts/format.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+cargo fmt
