--- test.track.json
@@ -1,1 +1,0 @@
-{"label":"","creator":"","description":"","duration":0,"version":"6.2","startPosition":{"x":0.0,"y":0.0},"lines":[{"id":1,"type":1,"x1":-20.88749649077696,"y1":-13.79999830126782,"x2":-8.043748021871089,"y2":-15.956248044222821,"flipped":false,"leftExtended":false,"rightExtended":false,"multiplier":7.0},{"id":2,"type":0,"x1":24.956248044222818,"y1":-7.706249027699343,"x2":41.64374605491804,"y2":-9.112498860061297,"flipped":false,"leftExtended":false,"rightExtended":false},{"id":0,"type":2,"x1":1.0500008940695622,"y1":-23.643747127801518,"x2":23.737498189509125,"y2":-22.799997228384342,"width":2.0}],"layers":[],"riders":[],"script":""}
