+++ tests/conversion.rs
@@ -1,7 +1,5 @@
 #[cfg(test)]
 mod tests {
-    // use lr_formatter_rs::{Format, convert};
-
     #[test]
     fn parse_json() {
         todo!("Test not implemented");
@@ -19,6 +17,36 @@ mod tests {
 
     #[test]
     fn write_lrb() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn parse_sol() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn write_sol() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn parse_lra_json() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn parse_trk() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn parse_boshtf() {
+        todo!("Test not implemented");
+    }
+
+    #[test]
+    fn parse_lrpk() {
         todo!("Test not implemented");
     }
 }
